membujuri pencarian dan akun rekam medis

+++ resources/views/pasien/pasien.blade.php
@@ -5,10 +5,7 @@
     <div>
         <form action="{{route('pasien/pasien')}}" method="get" class="row g-12">
             <div class="col-md-4">
-                <input class="form-control" type="text" name="nama"placeholder="Cari nama pasien" aria-label="default input example">
-            </div>
-            <div class="col-md-4">
-                <input class="form-control" type="text" name="no_berobat"placeholder="Cari pasien no berobat" aria-label="default input example">
+                <input class="form-control" type="text" name="cari"placeholder="Cari pasien" aria-label="default input example">
             </div>
             <div class="col-auto">
             <button type="submit" class="btn btn-primary mb-3"><i class="fa-sharp fa-solid fa-magnifying-glass"></i></button>
@@ -26,7 +23,7 @@
                 <th scope="col">No</th>
                 <th scope="col">No berobat</th>
                 <th scope="col">NIK</th>
-                <th scope="col">Jenis beobat</th>
+                <th scope="col">Jenis berobat</th>
                 <th scope="col">Nama</th>
                 <th scope="col">Tanggal lahir</th>
                 <th scope="col">tempat</th>
@@ -47,15 +44,20 @@
                     <td data-title="NIK" align="left">{{$pas->nik}}</td>
                     <td data-title="Jenis berobat">{{$pas->jenis_berobat}}</td>
                     <td data-title="Nama" style="text-transform: uppercase">{{$pas->nama}}</td>
-                    <td data-title="Tanggal lahir">{{$pas->tanggal}}</td>
+                    <td data-title="Tanggal lahir">{{date('d-m-Y',strtotime($pas->tanggal))}}</td>
                     <td data-title="Tempat lahir">{{$pas->tempat}}</td>
                     <td data-title="Alamat">{{$pas->alamat}}</td>
                     <td data-title="Gol darah">{{$pas->gol_darah}}</td>
                     <td data-title="No Hp">{{$pas->no_hp}}</td>
                     <td data-title="Aksi">
-                        <a href="detail/id={{$pas->id_pasien}}&pasien_id={{$pas->id_pasien}}" class="btn btn-success"><i class="fa-solid fa-book-medical"></i></i></a>
-                        <a href="edit_pasien/{{$pas->id_pasien}}" class="btn btn-warning"><i class="fa-solid fa-pen-to-square"></i></a>
-                        <a href="hapus_pasien/{{$pas->id_pasien}}" class="btn btn-danger" onclick="javascript: return confirm('Konfirmasi data akan dihapus');"><i class="fa-solid fa-trash"></i></a>
+                        @if(Auth::user()->level =='rekam_medis')
+                            <a href="detail/id={{$pas->id_pasien}}&pasien_id={{$pas->id_pasien}}" class="btn btn-success"><i class="fa-solid fa-book-medical"></i></i></a>
+                        @endif
+                        @if(Auth::user()->level =='admin')
+                            <a href="detail/id={{$pas->id_pasien}}&pasien_id={{$pas->id_pasien}}" class="btn btn-success"><i class="fa-solid fa-book-medical"></i></i></a>
+                            <a href="edit_pasien/{{$pas->id_pasien}}" class="btn btn-warning"><i class="fa-solid fa-pen-to-square"></i></a>
+                            <a href="hapus_pasien/{{$pas->id_pasien}}" class="btn btn-danger" onclick="javascript: return confirm('Konfirmasi data akan dihapus');"><i class="fa-solid fa-trash"></i></a>
+                        @endif
                     </td>
                 </tr>
             @endforeach
